refactor: use node --import tsx to run vite in client app
