+++ apps/client/package.json
@@ -4,9 +4,10 @@
   "type": "module",
   "private": true,
   "scripts": {
-    "dev": "vite --host 0.0.0.0",
-    "build": "vite build",
-    "preview": "vite preview"
+    "dev": "pnpm tsx-vite --host 0.0.0.0",
+    "build": "pnpm tsx-vite build",
+    "preview": "pnpm tsx-vite preview",
+    "tsx-vite": "node --import tsx ./node_modules/vite/bin/vite.js"
   },
   "dependencies": {
     "@mp/env": "workspace:*",
